Refactoring parse_elf into many files in elf module
