+++ ttexalens/elf/__init__.py
@@ -1,0 +1,19 @@
+# SPDX-FileCopyrightText: © 2025 Tenstorrent AI ULC
+
+# SPDX-License-Identifier: Apache-2.0
+
+from .cu import ElfCompileUnit
+from .die import ElfDie
+from .dwarf import ElfDwarf
+from .parsed import ParsedElfFile, ParsedElfFileWithOffset, read_elf
+from .variable import ElfVariable
+
+__all__ = [
+    "ElfCompileUnit",
+    "ElfDie",
+    "ElfDwarf",
+    "ElfVariable",
+    "ParsedElfFile",
+    "ParsedElfFileWithOffset",
+    "read_elf",
+]
